adds refactoring and styles fixes
adds HomeView which replaces MainPage component
adds refactoring of TestPage and fixes styles errors
