+++ src/components/TestPage/TestPage.module.css
@@ -1,8 +1,10 @@
 .questionForm {
   margin: 28px auto 20px;
+  padding: 50px 85px 58px 85px;
+
   max-width: 820px;
   min-height: 500px;
-  padding: 50px 85px 58px 85px;
+
   background-color: var(--white);
   box-shadow: 0px 4px 15px rgba(2, 23, 42, 0.1);
 }
@@ -80,7 +82,6 @@
 
 .optionBox {
   display: flex;
-  /* align-items: center; */
   align-items: center;
 }
 
@@ -101,13 +102,6 @@
   margin: -1px;
   padding: 0;
   border: 0;
-}
-
-.optionAnswer {
-  position: relative;
-  display: block;
-
-  padding-left: 38px;
 }
 
 .checkbox {
@@ -134,8 +128,17 @@
   border-radius: 50%;
 }
 
+.optionAnswer {
+  position: relative;
+  display: block;
+  padding-left: 38px;
+
+  overflow-wrap: anywhere;
+}
+
 .optionLabel {
   display: flex;
+
   font-weight: 500;
   font-size: 16px;
   line-height: 1.25;
@@ -147,140 +150,4 @@
   .optionLabel {
     font-size: 10px;
   }
-}
-
-.optionInput {
-  /* appearance: none; */
-  margin-right: 20px;
-}
-
-/* .checkbox {
-  display: inline-block;
-  width: 18px;
-  height: 18px;
-  border: 1px solid var(--black);
-  border-radius: 50%;
-  background-color: var(--white);
-}
-
-.optionInput:checked + .checkbox {
-  background-color: var(--accentColor);
-} */
-
-.buttonWrapper {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  margin-right: 210px;
-  margin-left: 210px;
-  padding-bottom: 93px;
-}
-
-@media screen and (max-width: 767px) {
-  .buttonWrapper {
-    margin: 0;
-    padding-bottom: 20px;
-  }
-}
-
-@media screen and (max-width: 1279px) {
-  .buttonWrapper {
-    margin-left: 0;
-    margin-right: 0;
-  }
-}
-
-.buttonPrevious,
-.buttonNext {
-  display: flex;
-  width: 210px;
-  border: none;
-  align-items: center;
-  padding-top: 12px;
-  padding-bottom: 12px;
-  font-weight: 500;
-  font-size: 12px;
-  line-height: 1.25;
-  background-color: var(--white);
-  color: var(--black);
-  fill: var(--black);
-  transition: all 150ms ease-in;
-  outline: none;
-}
-
-.buttonPrevious:hover,
-.buttonNext:hover,
-.buttonPrevious:focus,
-.buttonNext:focus {
-  background-color: var(--accentColor);
-  color: var(--white);
-  fill: var(--white);
-}
-
-.disabledPrevBtn {
-  composes: buttonPrevious;
-  pointer-events: none;
-  background-color: var(--secondaryTextColor);
-  color: var(--white);
-  fill: var(--white);
-}
-
-.disabledNextBtn {
-  composes: buttonNext;
-  pointer-events: none;
-  background-color: var(--labelTextColor);
-}
-
-.disabledPrevBtn:focus,
-.disabledNextBtn:focus {
-  background-color: var(--labelTextColor);
-  color: var(--black);
-}
-
-@media screen and (max-width: 767px) {
-  .buttonPrevious,
-  .buttonNext {
-    width: 55px;
-  }
-}
-
-.buttonPrevious {
-  padding-left: 15px;
-}
-
-@media screen and (max-width: 767px) {
-  .buttonPreviousName {
-    display: none;
-  }
-}
-
-.arrowPrevious {
-  margin-right: 11px;
-  transform: rotate(180deg);
-  fill: inherit;
-}
-
-.arrowBlack {
-  fill: inherit;
-}
-
-.buttonNext {
-  display: flex;
-  justify-content: space-between;
-  padding-right: 15px;
-  padding-left: 20px;
-  text-decoration: none;
-}
-
-@media screen and (max-width: 767px) {
-  .buttonNextName {
-    display: none;
-  }
-}
-
-.buttonSubmit {
-  composes: buttonNext;
-  display: block;
-
-  text-align: center;
 }
