+++ src/views/HomeView/HomeView.module.css
@@ -1,30 +1,3 @@
-.mainContainer {
-  margin-left: auto;
-  margin-right: auto;
-  padding-left: 20px;
-  padding-right: 20px;
-  max-width: 1280px;
-  background-color: #f5f6fb;
-  font-family: var(--mainFont);
-}
-@media screen and (min-width: var(--mobile)) {
-  .mainContainer {
-    width: var(--mobile);
-  }
-}
-
-@media screen and (min-width: var(--tablet)) {
-  .mainContainer {
-    width: var(--tablet);
-  }
-}
-
-@media screen and (min-width: var(--desktop)) {
-  .mainContainer {
-    width: var(--desktop);
-  }
-}
-
 .quote {
   padding-top: 122px;
   margin: 0;
